# update goshuirc/ircmsg to latest master