… vendor/github.com/goshuirc/irc-go/ircmsg/message.go
… vendor/github.com/goshuirc/irc-go/ircmsg/tags.go