Thanks for these benchmarks, they were quite helpful!

On my system (a low-end Kaby Lake without AVX) we are now losing to you by a factor of at most 2 :-)

````
goos: linux
goarch: amd64
pkg: github.com/jakebailey/irc-benchmarks
BenchmarkEncodeSimple/jakebailey/irc-4         	20000000	       108 ns/op
BenchmarkEncodeSimple/jakebailey/irc_WriteTo-4 	20000000	        92.1 ns/op
BenchmarkEncodeSimple/jakebailey/ircold-4      	20000000	        94.3 ns/op
BenchmarkEncodeSimple/sorcix/irc-4             	20000000	        94.8 ns/op
BenchmarkEncodeSimple/goshuirc/irc-go/ircmsg-4 	20000000	        96.3 ns/op
BenchmarkEncodeTwitch/jakebailey/irc-4         	 2000000	       782 ns/op
BenchmarkEncodeTwitch/jakebailey/irc_WriteTo-4 	 2000000	       695 ns/op
BenchmarkEncodeTwitch/jakebailey/ircold-4      	 1000000	      1110 ns/op
BenchmarkEncodeTwitch/sorcix/irc-4             	 1000000	      1108 ns/op
BenchmarkEncodeTwitch/goshuirc/irc-go/ircmsg-4 	 1000000	      1395 ns/op
BenchmarkEncodeEscaping/jakebailey/irc-4       	 1000000	      1290 ns/op
BenchmarkEncodeEscaping/jakebailey/irc_WriteTo-4         	 1000000	      1145 ns/op
BenchmarkEncodeEscaping/jakebailey/ircold-4              	 1000000	      1242 ns/op
BenchmarkEncodeEscaping/sorcix/irc-4                     	 1000000	      1261 ns/op
BenchmarkEncodeEscaping/goshuirc/irc-go/ircmsg-4         	 1000000	      2042 ns/op
BenchmarkParseSimple/jakebailey/irc-4                    	20000000	       121 ns/op
BenchmarkParseSimple/jakebailey/ircold-4                 	10000000	       193 ns/op
BenchmarkParseSimple/fluffle/goirc/client-4              	 5000000	       331 ns/op
BenchmarkParseSimple/sorcix/irc-4                        	10000000	       188 ns/op
BenchmarkParseSimple/thoj/go-ircevent-4                  	 5000000	       312 ns/op
BenchmarkParseSimple/goshuirc/irc-go/ircmsg-4            	10000000	       148 ns/op
BenchmarkParseSimple/gempir/go-twitch-irc-4              	10000000	       176 ns/op
BenchmarkParseTwitch/jakebailey/irc-4                    	 1000000	      1173 ns/op
BenchmarkParseTwitch/jakebailey/ircold-4                 	  300000	      5251 ns/op
BenchmarkParseTwitch/fluffle/goirc/client-4              	  300000	      5417 ns/op
BenchmarkParseTwitch/sorcix/irc-4                        	  300000	      5224 ns/op
BenchmarkParseTwitch/thoj/go-ircevent-4                  	  300000	      4210 ns/op
BenchmarkParseTwitch/goshuirc/irc-go/ircmsg-4            	 1000000	      2080 ns/op
BenchmarkParseTwitch/gempir/go-twitch-irc-4              	  300000	      4174 ns/op
BenchmarkParseEscaping/jakebailey/irc-4                  	 1000000	      2312 ns/op
BenchmarkParseEscaping/jakebailey/ircold-4               	  200000	      7231 ns/op
BenchmarkParseEscaping/fluffle/goirc/client-4            	  200000	      7363 ns/op
BenchmarkParseEscaping/sorcix/irc-4                      	  200000	      7274 ns/op
BenchmarkParseEscaping/thoj/go-ircevent-4                	  300000	      5895 ns/op
BenchmarkParseEscaping/goshuirc/irc-go/ircmsg-4          	  500000	      2906 ns/op
BenchmarkParseEscaping/gempir/go-twitch-irc-4            	  300000	      4951 ns/op
PASS
ok  	github.com/jakebailey/irc-benchmarks	62.210s
````

Incidentally, how do you get the suite to display memory statistics? `go test -bench -benchmem .` doesn't run the benchmark suite at all for me (I'm on golang 1.12).