@jakebailey 2019-03-07
Thanks for the PR, I'll have to see if I want to merge this or do something myself to update all of these at once (it's been a while).

`-bench` is the flag which controls which benchmarks to run. You have to do `go test -bench . -benchmem`. Better yet would be for me to call `b.ReportAllocs()` in each of the benchmarks to force that info to be printed.

Shame the performance boost came at the cost of a breaking change; good boost in any case:

```
name                                     old time/op    new time/op    delta
EncodeSimple/goshuirc/irc-go/ircmsg-4       116ns ± 0%     105ns ± 1%   -9.83%  (p=0.000 n=7+10)
EncodeTwitch/goshuirc/irc-go/ircmsg-4      1.61µs ± 1%    1.51µs ± 2%   -6.17%  (p=0.000 n=8+10)
EncodeEscaping/goshuirc/irc-go/ircmsg-4    2.32µs ± 1%    2.19µs ± 1%   -5.40%  (p=0.000 n=10+10)
ParseSimple/goshuirc/irc-go/ircmsg-4        628ns ± 1%     162ns ± 0%  -74.13%  (p=0.000 n=9+8)
ParseTwitch/goshuirc/irc-go/ircmsg-4       10.9µs ± 0%     2.2µs ± 1%  -79.59%  (p=0.000 n=10+10)
ParseEscaping/goshuirc/irc-go/ircmsg-4     17.2µs ± 0%     3.1µs ± 1%  -82.19%  (p=0.000 n=10+10)

name                                     old alloc/op   new alloc/op   delta
EncodeSimple/goshuirc/irc-go/ircmsg-4       64.0B ± 0%     64.0B ± 0%     ~     (all equal)
EncodeTwitch/goshuirc/irc-go/ircmsg-4      1.19kB ± 3%    1.18kB ± 5%     ~     (p=0.108 n=9+10)
EncodeEscaping/goshuirc/irc-go/ircmsg-4    1.61kB ± 6%    1.59kB ± 5%     ~     (p=0.315 n=10+10)
ParseSimple/goshuirc/irc-go/ircmsg-4         336B ± 0%       48B ± 0%  -85.71%  (p=0.000 n=10+10)
ParseTwitch/goshuirc/irc-go/ircmsg-4       5.87kB ± 0%    2.19kB ± 0%  -62.69%  (p=0.000 n=10+10)
ParseEscaping/goshuirc/irc-go/ircmsg-4     10.2kB ± 0%     2.5kB ± 0%  -75.94%  (p=0.000 n=10+10)

name                                     old allocs/op  new allocs/op  delta
EncodeSimple/goshuirc/irc-go/ircmsg-4        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
EncodeTwitch/goshuirc/irc-go/ircmsg-4        3.70 ±19%      3.70 ±19%     ~     (p=1.000 n=10+10)
EncodeEscaping/goshuirc/irc-go/ircmsg-4      8.00 ± 0%      8.00 ± 0%     ~     (all equal)
ParseSimple/goshuirc/irc-go/ircmsg-4         11.0 ± 0%       2.0 ± 0%  -81.82%  (p=0.000 n=10+10)
ParseTwitch/goshuirc/irc-go/ircmsg-4          155 ± 0%         6 ± 0%  -96.13%  (p=0.000 n=10+10)
ParseEscaping/goshuirc/irc-go/ircmsg-4        246 ± 0%         9 ± 0%  -96.34%  (p=0.000 n=10+10)
```

I'll take a look at this later tonight.