@slingamn 2019-03-07
Ah, the API break is unrelated to the optimization, it's because we implemented this: https://ircv3.net/specs/extensions/message-tags.html