@jakebailey 2019-03-07
Didn't irc-go implement IRCv3 tags before? Or is that a different thing called "tags" in the spec?