@slingamn 2019-03-07
Previously we were implementing a draft version of the specification that didn't specify length limits on tag data. The ratified version of the specification (which just came out recently) has somewhat complicated requirements for tag lengths:

1. Clients can send at most 4094 bytes of tag data, including client-only tags with the `+` prefix
1. Servers can relay at most 4094 bytes of client-only tags, and then they get to add at most 4094 bytes of their own tag data (without the `+` prefix)
1. Servers must reply with `417 ERR_INPUTTOOLONG` if the client sends tag data over the limit