@jakebailey 2019-03-07
Ah, I see from the PR you made to the other repo; you changed a number of things to be more correct/spec compliant. I was wondering how you found this repo, likely because you were checking for users who you'd break. The lookup table is neat, I'll have to see if it's any faster than a switch case (though I'm not convinced it works when tags contain code points above 255, very common in Twitch tags like resub messages or non-ASCII display names).

My implementation isn't perfect and probably not compliant if messages get too long. My main use-case is for Twitch, which breaks rules anyway, so I do the minimum work and just 1:1 in and out with only a basic set of sanity checks.

I didn't realize that v3 was finally ratified. I'll have to read when I have time. Back in 2015 I went around to all of the Go IRC libs I could find to add tag support, and it was still new and really only interesting because Twitch had adopted it for its chat servers. I'm not sure if I'd want to change my implementation to be more spec compliant now that it's finalized. We'll see, I have other things I'm working on.