@slingamn 2019-03-07
Yeah, I did a Github code search for clients (we didn't find any besides @oragono, other internal projects, and this repo).

My understanding of the tag value escaping specification is that it is defined at the byte level, not the codepoint level (like other RFC1459-and-successor message syntax definitions), so unescaping individual bytes via a lookup table should be correct. (Escaping via `strings.Replacer` should be correct as well --- golang's `string` is a sequence of arbitrary bytes, and character encoding should only come into play when iterating over a string with `range`.)